Update contents.md
I rearranged because i felt like establishing trust (chapter 1) and building community (2) with great communication (3) are essential for being able to collaborate (4). 

I'm on the fence about moving Leading and Managing up – on the one hand, orientation-type knowledge is helpful for executing practical knowledge (e.g. if our audience is leaders, it might be useful for them to understand a bit of theory first). On the other hand, I believe that good leadership has a realistic view on power dynamics and how they're propagated...